GUI application
GUI which uses the QS "library" i've made. Added a way to choose a subject, which then makes the program get the queue for that subject and loads it into a listbox. So basically it can only view the QS queue as of now. Planning on implementing more features soon